+++ toys-core/CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ### 0.3.10 / TBD
 
+* CHANGED: Dropped Tool#option. Use Tool#get instead.
 * CHANGED: "run" directive renamed to "to_run"
 * CHANGED: Highline mixin now uses Highline 2.0
 * ADDED: Mixins can provide initializers
+* ADDED: Loader can load an inline block
 
 ### 0.3.9.1 / 2018-06-24